fix: Change all T2 id fields from i32 to i64 for SQLite compatibility
SQLite INTEGER type maps to i64 in Rust. This fixes the
"mismatched types; Rust type i64 is not compatible with SQL type INT4" error.

Changed all id fields and related parameters across:
- t2/models.rs - all struct fields
- t2/routes.rs - all route parameters and query results
- t2/guards.rs - auth guard types
- t2/ai.rs - AI response types

+++ server/src/t2/ai.rs
@@ -267,9 +267,9 @@ match_score от 0 до 100 - насколько товар соответств
 
     #[derive(Deserialize)]
     struct AIRecommendation {
-        product_id: i32,
+        product_id: i64,
         price_category: String,
-        match_score: i32,
+        match_score: i64,
         match_reasons: Vec<String>,
     }
 
@@ -366,7 +366,7 @@ pub async fn recommend_accessories(
 
     #[derive(Deserialize)]
     struct AIAccessory {
-        product_id: i32,
+        product_id: i64,
         reason: String,
         benefit: String,
     }
@@ -459,7 +459,7 @@ pub async fn recommend_tariffs(
 
     #[derive(Deserialize)]
     struct AITariff {
-        tariff_id: i32,
+        tariff_id: i64,
         recommendation: String,
     }
